1. Changed the APIs to refer to subjects as discussed last week. 2. Added the ability to assign globally unique schema id on the master and retrieve schemas using ids 3. Unit tests need to be modified
Conflicts:
	config/schema-registry.properties
	src/main/java/io/confluent/kafka/schemaregistry/rest/Main.java
	src/main/java/io/confluent/kafka/schemaregistry/rest/RegisterSchemaForwardingAgent.java
	src/main/java/io/confluent/kafka/schemaregistry/rest/SchemaRegistryConfig.java
	src/main/java/io/confluent/kafka/schemaregistry/rest/SchemaRegistryRestApplication.java
	src/main/java/io/confluent/kafka/schemaregistry/rest/entities/requests/RegisterSchemaResponse.java
	src/main/java/io/confluent/kafka/schemaregistry/rest/resources/SchemasResource.java
	src/main/java/io/confluent/kafka/schemaregistry/rest/resources/SubjectsResource.java
	src/main/java/io/confluent/kafka/schemaregistry/storage/KafkaSchemaRegistry.java
	src/main/java/io/confluent/kafka/schemaregistry/storage/KafkaStore.java
	src/main/java/io/confluent/kafka/schemaregistry/storage/SchemaRegistry.java
	src/main/java/io/confluent/kafka/schemaregistry/utils/RestUtils.java
	src/test/java/io/confluent/kafka/schemaregistry/RestApp.java
	src/test/java/io/confluent/kafka/schemaregistry/storage/StringMessageHandler.java
	src/test/java/io/confluent/kafka/schemaregistry/utils/TestUtils.java

+++ src/test/java/io/confluent/kafka/schemaregistry/utils/TestUtils.java
@@ -97,7 +97,7 @@ public class TestUtils {
     }
   }
 
-  public static int registerSchema(String baseUrl, String schemaString, String subject)
+  public static long registerSchema(String baseUrl, String schemaString, String subject)
       throws IOException {
     RegisterSchemaRequest request = new RegisterSchemaRequest();
     request.setSchema(schemaString);
